restructured boundary.py into a subpackage
