+++ src/pybella/flow_solver/physics/hydrostatics.py
@@ -1,7 +1,7 @@
 import numpy as np
 import numba as nb
 
-from ..utils import boundary as bdry
+from ..utils.boundary import node_boundary as bdry_n
 
 
 def column(HydroState, HydroState_n, Y, Y_n, elem, node, th, ud):
@@ -306,7 +306,7 @@ def initial_pressure(Sol, npf, elem, node, ud, th):
     npf.p2_cells[x_idx, y_idx] += pibot[x_idx].reshape(
         -1, 1
     ) - 1.0 * npf.HydroState.p20[y_idx].reshape(1, -1)
-    bdry.set_ghostcells_p2(npf.p2_cells, elem, ud)
+    bdry_n.set_ghostcells_p2(npf.p2_cells, elem, ud)
 
     icxn = node.icx
     icyn = node.icy
@@ -364,7 +364,7 @@ def initial_pressure(Sol, npf, elem, node, ud, th):
     sgn[1::2] *= -1
 
     npf.p2_nodes[igx:-igx, igy:-igy] += sgn * delp2
-    bdry.set_ghostnodes_p2(npf.p2_nodes, node, ud)
+    bdry_n.set_ghostnodes_p2(npf.p2_nodes, node, ud)
 
     npf.dp2_nodes[:, :] = 0.0
