--- src/pybella/flow_solver/utils/boundary.py
@@ -1,580 +1,0 @@
-"""
-For more details on this module, refer to the write-up :ref:`boundary_handling`.
-"""
-
-import copy
-import numpy as np
-from ...utils import options as opts
-from ...utils import io
-
-
-def set_explicit_boundary_data(Sol, elem, ud, th, npf, step=None):
-    """
-    In-place update of the ghost cells in :class:`management.variable.Vars` given the boundary conditions specified by :class:`inputs.user_data.UserDataInit`.
-
-    Parameters
-    ----------
-    Sol : :class:`management.variable.Vars`
-        Solution data container
-    elem : :class:`discretization.kgrid.ElemSpaceDiscr`
-        Cells grid
-    ud : :class:`inputs.user_data.UserDataInit`
-        Data container for the initial conditions
-    th : :class:`physics.gas_dynamics.thermodynamic.init`
-        Thermodynamic variables of the system
-    npf : :class:`physics.low_mach.npf.MPV`
-        Variables relating to the elliptic solver
-    step : int, optional
-        Current step
-
-    """
-    igs = elem.igs
-    ndim = elem.ndim
-
-    # if step parameter is not None, then we are in the advection directional Strang-splitting, where the array has already flipped, and we should only update the relevant boundaries, i.e. those in the direction of the current Strang-split-step.
-    if step == None:
-        dims = np.arange(ndim)
-    else:
-        dims = [ndim - 1]
-
-    for dim in dims:
-        if step is not None:
-            current_step = step
-        else:
-            current_step = dim
-        ghost_padding, idx = get_ghost_padding(ndim, dim, igs)
-
-        if ud.gravity_strength[current_step] == 0.0:
-            # Do this for the axes that do not have gravity.
-            # Periodic BC.
-            if ud.bdry_type[current_step] == opts.BdryType.PERIODIC:
-                set_boundary(Sol, ghost_padding, "wrap", idx, step=None)
-            # Wall BC.
-            elif ud.bdry_type[current_step] == opts.BdryType.WALL:
-                set_boundary(Sol, ghost_padding, "symmetric", idx, step=None)
-            elif ud.bdry_type[current_step] == opts.BdryType.RAYLEIGH:
-                assert 0, "Rayleigh boundary not defined on x-direction."
-
-        else:
-            # get current axis that has gravity.
-            gravity_axis = dim
-
-            direction = -1.0
-            offset = 0
-
-            # get gravity strength specified in the user data file.
-            g = ud.gravity_strength[gravity_axis]
-
-            # for the number of ghost cells in the gravity axis...
-            for side in ghost_padding[gravity_axis]:
-                direction *= -1
-                # loop through each of these ghost cells.
-                for current_idx in np.arange(side)[::-1]:
-                    if step != None:
-                        y_axs = ndim - 1
-                    else:
-                        y_axs = 1
-                    nlast, nsource, nimage = get_gravity_padding(
-                        ndim, current_idx, direction, offset, elem, y_axs=y_axs
-                    )
-
-                    Y_last = Sol.rhoY[nlast] / Sol.rho[nlast]
-
-                    rhoYv_image = (
-                        -Sol.rhov[nsource] * Sol.rhoY[nsource] / Sol.rho[nsource]
-                    )
-
-                    S = 1.0 / ud.stratification(elem.y[nimage[y_axs]])
-
-                    if hasattr(ud, "ATMOSPHERIC_EXTENSION"):
-                        dpi = (
-                            npf.HydroState.p20[nimage[y_axs]]
-                            - npf.HydroState.p20[nlast[y_axs]]
-                        ) * ud.Msq
-                    else:
-                        dpi = (
-                            direction
-                            * (th.Gamma * g)
-                            * 0.5
-                            * elem.dy
-                            * (1.0 / Y_last + S)
-                        )
-
-                    rhoY = (
-                        ((Sol.rhoY[nlast] ** th.gm1) + dpi) ** th.gm1inv
-                        if ud.is_compressible == 1
-                        else npf.HydroState.rhoY0[nimage[y_axs]]
-                    )
-
-                    rho = rhoY * S
-
-                    Y_source = Sol.rhoY[nsource] / Sol.rho[nsource]
-                    Y_image = rhoY / rho
-
-                    if hasattr(ud, "ATMOSPHERIC_EXTENSION"):
-                        if direction > 0:  # if bottom boundary
-                            v = Sol.rhov[nsource] * Y_source / Sol.rho[nsource] * rho
-                        else:  # if top boundary
-                            v = Sol.rhov[nsource] * Y_source
-
-                        Th_slc = rhoY / rho / Y_last
-
-                    else:
-                        v = rhoYv_image / rhoY
-                        Th_slc = 1.0
-
-                    u = Sol.rhou[nsource] / Sol.rho[nsource]
-                    w = Sol.rhow[nsource] / Sol.rho[nsource]
-                    X = Sol.rhoX[nsource] / Sol.rho[nsource]
-
-                    Sol.rho[nimage] = rho
-                    Sol.rhou[nimage] = rho * u * Th_slc
-                    if hasattr(ud, "ATMOSPHERIC_EXTENSION"):
-                        Sol.rhov[nimage] = -v / Y_image
-                    else:
-                        Sol.rhov[nimage] = rho * v
-                    Sol.rhow[nimage] = rho * w * Th_slc
-                    Sol.rhoY[nimage] = rhoY
-                    Sol.rhoX[nimage] = rho * X
-
-                offset += 1
-
-
-def set_boundary(Sol, pads, btype, idx, step=None):
-    """
-    Called by the function :func:`inputs.boundary.set_explicit_boundary_data`. Pads in-place the ghost cells for a given boundary type.
-
-    Parameters
-    ----------
-    Sol : :class:`management.variable.Vars`
-        Solution data container.
-    pads : tuple
-        A tuple containing the number of ghost cells to pad at each end.
-    btype : string
-        The type of boundary condition to pad. Currently supports:
-            * `wrap` for periodic boundary conditions
-            * `symmetric` for wall boundary conditions
-            * `negative_symmetric` for wall boundary conditions, but with the signs flipped.
-    idx : tuple
-        A tuple containing the slice indices for the inner array, e.g. `(slice(2,-2),slice(2,-2))` for a 2D-array with 2 ghost cells for all edges.
-    step : int, optional
-        If we are in the advection routine with the flipped arrays according to the directional Strang-splitting, we want to pad the correct direction. `step=0`, pads the x-direction while `step=1` pads the y-direction.
-
-    """
-    Sol.rho[...] = np.pad(Sol.rho[idx], pads, btype)
-
-    if btype == "symmetric":
-        Sol.rhov[...] = np.pad(Sol.rhov[idx], pads, negative_symmetric)
-        Sol.rho[...] = np.pad(Sol.rho[idx], pads, "symmetric")
-        Sol.rhou[...] = np.pad(Sol.rhou[idx], pads, "symmetric")
-    elif btype == "constant":
-        Sol.rho[...] = np.pad(Sol.rho[idx], pads, "symmetric")
-        Sol.rhou[...] = np.pad(Sol.rhou[idx], pads, "symmetric")
-        Sol.rhov[...] = np.pad(Sol.rhov[idx], pads, btype)
-        Sol.rhow[...] = np.pad(Sol.rhow[idx], pads, "symmetric")
-        btype = "symmetric"
-    else:
-        Sol.rhou[...] = np.pad(Sol.rhou[idx], pads, btype)
-        Sol.rhov[...] = np.pad(Sol.rhov[idx], pads, btype)
-        Sol.rhow[...] = np.pad(Sol.rhow[idx], pads, btype)
-
-    Sol.rhoY[...] = np.pad(Sol.rhoY[idx], pads, btype)
-    Sol.rhoX[...] = np.pad(Sol.rhoX[idx], pads, btype)
-
-
-def negative_symmetric(vector, pad_width, iaxis, kwargs=None):
-    """
-    Taken from the reference:
-
-    Parameters
-    ----------
-    vector : ndarray
-        A rank 1 array already padded with zeros. Padded values are vector `[:iaxis_pad_width[0]] and vector[-iaxis_pad_width[1]:]`.
-    iaxis_pad_width : tuple
-        A 2-tuple of ints, `iaxis_pad_width[0]` represents the number of values padded at the beginning of vector where `iaxis_pad_width[1]` represents the number of values padded at the end of vector.
-    iaxis : int
-        The axis currently being calculated.
-    kwargs : dict
-        Any keyword arguments the function requires.
-
-    References
-    ----------
-    https://docs.scipy.org/doc/numpy/reference/generated/numpy.pad.html
-
-    """
-    if pad_width[1] > 0:
-        sign = -1
-        vector[: pad_width[0]] = sign * vector[pad_width[0] : 2 * pad_width[0]][::-1]
-        vector[-pad_width[1] :] = sign * vector[-2 * pad_width[1] : -pad_width[1]][::-1]
-        return vector
-    else:  # axis must have length > 0 for padding
-        return vector
-
-
-def get_gravity_padding(ndim, cur_idx, direction, offset, elem, y_axs=None):
-    """
-    Parameters
-    ----------
-    ndim : int
-        Number of dimensions.
-    cur_idx : int
-        The current index of the ghost cell in the gravity direction to be updated.
-    direction : int
-        Top of the domain, `direction=+1`, bottom of the domain, `direction=-1`.
-    offset : int
-        `offset=0`, index starts counting from 0,1.... `offset=1`, index starts counting from -1,-2,..., i.e. end-selection of the array.
-    elem : :class:`discretization.kgrid.ElemSpaceDiscr`
-        Cell grid.
-    y_axs : int, optional
-        `Default == None`. Specifies the direction of the gravity axis. If `None`, then direction is the the y-axis.
-
-    """
-    cur_i = np.copy(cur_idx)
-    cur_idx += offset * ((elem.icy - 1) - 2 * cur_idx)
-    gravity_padding = [slice(None)] * ndim
-    if y_axs == None:
-        # y_axs = ndim - 1
-        y_axs = 1
-
-    nlast = np.copy(gravity_padding)
-    nlast[y_axs] = int(cur_idx + direction)
-
-    nsource = np.copy(gravity_padding)
-    nsource[y_axs] = int(
-        offset * (elem.icy) + direction * (2 * elem.igy - (1 - offset) - cur_i)
-    )
-
-    nimage = np.copy(gravity_padding)
-    nimage[y_axs] = int(cur_idx)
-    return tuple(nlast), tuple(nsource), tuple(nimage)
-
-
-def set_ghostcells_p2(p, elem, ud):
-    igs = elem.igs
-
-    for dim in range(elem.ndim):
-        ghost_padding, idx = get_ghost_padding(elem.ndim, dim, igs)
-
-        if ud.bdry_type[dim] == opts.BdryType.PERIODIC:
-            p[...] = np.pad(p[idx], ghost_padding, "wrap")
-        else:  # WALL
-            p[...] = np.pad(p[idx], ghost_padding, "symmetric")
-
-
-def set_ghostnodes_p2(p, node, ud, igs=None):
-    if igs is None:
-        igs = node.igs
-    for dim in range(node.ndim):
-        ghost_padding, idx = get_ghost_padding(node.ndim, dim, igs)
-
-        if ud.bdry_type[dim] == opts.BdryType.PERIODIC:
-            p[...] = np.pad(p[idx], ghost_padding, periodic_plus_one)
-        else:  # ud.bdry_type[dim] == opts.BdryType.WALL:
-            p[...] = np.pad(p[idx], ghost_padding, "reflect")
-
-    # if periodic_plus_one
-    if node.iicy == 2:  # implying horizontal slices
-        pn = p[:, 2, :]
-        pn = np.expand_dims(pn, axis=1)
-        p[...] = np.repeat(pn, node.icy, axis=1)
-
-
-def get_ghost_padding(ndim, dim, igs):
-    """
-    For a given direction, return the number of ghost cells to pad the current direction, and the index slice of the inner array without the ghost cells.
-
-    Parameters
-    ----------
-    ndim : int
-        Number of dimensions for the problem.
-    dim : int
-        Current dimension to update
-    igs : list
-        A list of number of ghost cells in all dimensions, e.g. `[2,2,2]` for 2 ghost cells in the x, y, and z directions.
-
-    Returns
-    -------
-    tuple
-        Number of ghost cells in the current dimension at both edges.
-    tuple
-        Index slice of the inner domain of the array.
-
-    """
-    ghost_padding = [(0, 0)] * ndim
-    ghost_padding[dim] = (igs[dim], igs[dim])
-
-    padded_idx = np.empty((ndim), dtype=object)
-    for idim in range(ndim):
-        padded_idx[idim] = slice(igs[idim], -igs[idim])
-    padded_idx[dim] = slice(None)
-
-    inner_domain = [slice(None)] * ndim
-    inner_domain[dim] = slice(igs[dim], -igs[dim])
-
-    return tuple(ghost_padding), tuple(inner_domain)
-
-
-def periodic_plus_one(vector, pad_width, iaxis, kwargs=None):
-    """
-    Taken from the reference:
-
-    Parameters
-    ----------
-    vector : ndarray
-        A rank 1 array already padded with zeros. Padded values are vector `[:iaxis_pad_width[0]] and vector[-iaxis_pad_width[1]:]`.
-    iaxis_pad_width : tuple
-        A 2-tuple of ints, `iaxis_pad_width[0]` represents the number of values padded at the beginning of vector where `iaxis_pad_width[1]` represents the number of values padded at the end of vector.
-    iaxis : int
-        The axis currently being calculated.
-    kwargs : dict
-        Any keyword arguments the function requires.
-    References
-    ----------
-    https://docs.scipy.org/doc/numpy/reference/generated/numpy.pad.html
-
-    """
-    if all(pad_width) > 0:
-        vector[: pad_width[0] + 1], vector[-pad_width[1] - 1 :] = (
-            vector[-pad_width[1] - pad_width[1] - 1 : -pad_width[1]],
-            vector[pad_width[0] : pad_width[0] + pad_width[0] + 1].copy(),
-        )
-    return vector
-
-
-def get_tau_y(ud, elem, node, alpha):
-    tauc_y = np.zeros_like(elem.y)
-    taun_y = np.zeros_like(node.y)
-
-    # ud.bcy = elem.y[-ud.inbcy-2]
-    ud.bny = node.y[-ud.inbcy - 3]
-    # ud.bny = ud.bcy
-
-    # c1c = elem.y <= ud.bcy
-    # ccc = (elem.y[:-2] - ud.bcy) / (elem.y[:-2][-1] - ud.bcy)
-    # c2c = np.logical_and(ccc >= 0.0, ccc <= 0.5)
-    # c3c = np.logical_and(ccc > 0.5, ccc <= 1.0)
-
-    c1n = node.y <= ud.bny
-    ccn = (node.y[:-2] - ud.bny) / (node.y[:-2][-1] - ud.bny)
-    c2n = np.logical_and(ccn >= 0.0, ccn <= 0.5)
-    c3n = np.logical_and(ccn > 0.5, ccn <= 1.0)
-
-    # ccc and ccn can be reused below
-
-    # tauc_y[np.where(c1c)] = 0.0
-    # tauc_y[np.where(c2c)] = - alpha / 2.0 * (1.0 - np.cos( (elem.y[np.where(c2c)] - ud.bcy) / (elem.y[-1] - ud.bcy) * np.pi ))
-    # tauc_y[np.where(c3c)] = - alpha / 2.0 * (1.0 + ((elem.y[np.where(c3c)] - ud.bcy) / (elem.y[-1] - ud.bcy) - 0.5) * np.pi)
-
-    taun_y[np.where(c1n)] = 0.0
-    taun_y[np.where(c2n)] = (
-        -alpha
-        / 2.0
-        * (
-            1.0
-            - np.cos((node.y[np.where(c2n)] - ud.bny) / (node.y[-1] - ud.bny) * np.pi)
-        )
-    )
-    taun_y[np.where(c3n)] = (
-        -alpha
-        / 2.0
-        * (
-            1.0
-            + ((node.y[np.where(c3n)] - ud.bny) / (node.y[-1] - ud.bny) - 0.5) * np.pi
-        )
-    )
-
-    taun_y[-2:] = -np.abs(taun_y).max()
-    tauc_y[...] = np.interp(elem.y, node.y, taun_y)
-    # dd = 1.0
-    # tauc_y = dd * tauc_y / np.abs(tauc_y).max()
-    # taun_y = dd * taun_y / np.abs(taun_y).max()
-    return tauc_y, taun_y
-
-
-def get_bottom_tau_y(ud, elem, node, alpha, cutoff=0.5):
-    tauc_y = np.zeros_like(elem.y)
-    taun_y = np.zeros_like(node.y)
-
-    assert ud.ymax > cutoff, "rayleigh forcing boundary below minimum domain extent"
-    idx = (np.abs(elem.y - (ud.ymax - cutoff))).argmin()
-
-    ud.forcing_bny = node.y[idx]
-
-    c1n = node.y <= ud.forcing_bny
-    ccn = (node.y[:-3] - ud.forcing_bny) / (node.y[:-3][-1] - ud.forcing_bny)
-    c2n = np.logical_and(ccn >= 0.0, ccn <= 0.5)
-    c3n = np.logical_and(ccn > 0.5, ccn <= 1.0)
-
-    taun_y[np.where(c1n)] = 0.0
-    taun_y[np.where(c2n)] = (
-        -alpha
-        / 2.0
-        * (
-            1.0
-            - np.cos(
-                (node.y[np.where(c2n)] - ud.forcing_bny)
-                / (node.y[-1] - ud.forcing_bny)
-                * np.pi
-            )
-        )
-    )
-    taun_y[np.where(c3n)] = (
-        -alpha
-        / 2.0
-        * (
-            1.0
-            + (
-                (node.y[np.where(c3n)] - ud.forcing_bny) / (node.y[-1] - ud.forcing_bny)
-                - 0.5
-            )
-            * np.pi
-        )
-    )
-
-    taun_y[-3:] = -np.abs(taun_y).max()
-    taun_y[...] = taun_y[::-1]
-    tauc_y = np.interp(elem.y, node.y, taun_y)
-
-    dd = 1.0
-    tauc_y = dd * tauc_y / np.abs(tauc_y).max()
-    taun_y = dd * taun_y / np.abs(taun_y).max()
-
-    # tauc_y = tauc_y[2:-2]
-    # taun_y = taun_y[2:-2]
-
-    return tauc_y, taun_y
-
-
-def apply_rayleigh_forcing(
-    Sol,
-    npf,
-    ud,
-    elem,
-    node,
-    t,
-    step,
-    dt,
-    th,
-    bdry,
-    half=True,
-    Sol_half_new=None,
-    npf_half_new=None,
-):
-    """Apply Rayleigh forcing boundary condition (file or function based)."""
-    if not (hasattr(ud, "rayleigh_forcing") and ud.rayleigh_forcing):
-        return
-
-    t_offset = 0.5 * dt if half else dt
-
-    if ud.rayleigh_forcing_type == "file":
-        reader = io.read_input(ud.rayleigh_forcing_fn, ud.rayleigh_forcing_path)
-
-        if Sol_half_new is None or npf_half_new is None:
-            Sol_half_new = copy.deepcopy(Sol)
-            npf_half_new = copy.deepcopy(npf)
-
-        time_tag = "%.3d_after_full_step" % step
-        reader.get_data(Sol_half_new, npf_half_new, time_tag, half=half)
-
-        up = Sol_half_new.rhou / Sol_half_new.rho
-        vp = Sol_half_new.rhov / Sol_half_new.rho
-        Yp = Sol_half_new.rhoY / Sol_half_new.rho - npf.HydroState.Y0.reshape(1, -1)
-        pi = npf_half_new.p2_nodes
-
-        bdry.rayleigh_damping(Sol, npf, ud, elem, node, [up, vp, Yp, pi, t + t_offset])
-
-    elif ud.rayleigh_forcing_type == "func":
-        s = 5.0e-3 + 1e-4 + 0e-5
-        ud.rf_bot.eigenfunction(t + t_offset, s)
-        up, vp, Yp, pi = ud.rf_bot.dehatter(th)
-
-        ud.rf_bot.eigenfunction(t + t_offset, s, grid="n")
-        _, _, _, pi_n = ud.rf_bot.dehatter(th, grid="n")
-
-        bdry.rayleigh_damping(
-            Sol, npf, ud, elem, node, [up, vp, Yp, pi_n, t + t_offset]
-        )
-
-    bdry.set_explicit_boundary_data(Sol, elem, ud, th, npf)
-
-
-def rayleigh_damping(Sol, npf, ud, elem, node, forcing=None):
-    u = Sol.rhou / Sol.rho  # [elem.i2]
-    v = Sol.rhov / Sol.rho  # [elem.i2]
-    Y = Sol.rhoY / Sol.rho  # [elem.i2]
-    rho = Sol.rho  # [elem.i2]
-
-    if ud.bdry_type[1] == opts.BdryType.RAYLEIGH:
-        tcy, tny = ud.tcy, ud.tny
-    else:
-        tcy, tny = 0.0, 0.0
-
-    if forcing is not None:
-        tcy_f, tny_f = ud.forcing_tcy, ud.forcing_tny
-        tcy, tny = 0.0, 0.0
-
-        u_f, v_f, Y_f, pi_f, t = forcing
-
-        if ud.rayleigh_forcing_type == "file":
-            G = np.sqrt(9.0 / 40.0)
-            N = np.sqrt(ud.Nsq_ref)
-            C = ud.Cs * ud.u_ref
-            Gam = N * G / C
-            Om = ud.coriolis_strength[2] / 2.0 / ud.t_ref
-            # Om = 7.292 * 1e-5 / ud.t_ref
-            growth_rate = np.sqrt(Om * C * Gam)
-            mfac = np.exp(growth_rate * t * ud.t_ref)
-
-            # if np.all(ud.coriolis_strength) == 0.0:
-            # if ud.trad_forcing:
-            # mfac = 1.0
-        else:
-            mfac = 1.0
-
-        npf.p2_nodes[...] += tny_f * (npf.p2_nodes) + np.abs(tny_f) * mfac * pi_f
-        c_f = 1.0
-
-    else:
-        u_f, v_f, Y_f = 0.0, 0.0, 0.0
-        c_f = 0.0
-        tcy_f, tny_f = 0.0, 0.0
-        mfac = 0.0
-
-    # assuming 2D vertical slice - not dimension agnostic
-    u += tcy * (u - ud.u_wind_speed) + c_f * (
-        tcy_f * (u - ud.u_wind_speed) + np.abs(tcy_f) * mfac * u_f
-    )
-    v += tcy * (v - ud.v_wind_speed) + c_f * (
-        tcy_f * (v - ud.v_wind_speed) + np.abs(tcy_f) * mfac * v_f
-    )
-
-    Ybar = npf.HydroState.Y0.reshape(1, -1)
-    Y += tcy * (Y - Ybar) + c_f * (tcy_f * (Y - Ybar) + np.abs(tcy_f) * mfac * Y_f)
-
-    Sol.rhou[...] = rho * u
-    Sol.rhov[...] = rho * v
-    Sol.rhoY[...] = rho * Y
-
-def scale_wall_node_values(rhs, node, ud, factor=0.5):
-    """Scale values at wall boundary nodes by a given factor."""
-    ndim = node.ndim
-    igs = node.igs
-
-    for dim in range(ndim):
-        # Check if this dimension has wall boundaries
-        is_wall = (
-            ud.bdry_type[dim] == opts.BdryType.WALL
-            or ud.bdry_type[dim] == opts.BdryType.RAYLEIGH
-        )
-
-        if is_wall:
-            # Create index for all dimensions
-            idx = [slice(igs[d], -igs[d]) for d in range(ndim)]
-
-            # Scale first and last interior nodes in this dimension
-            for boundary_idx in [igs[dim], -igs[dim] - 1]:
-                idx[dim] = boundary_idx
-                rhs[tuple(idx)] *= factor
-                # rhs = rhs.at[tuple(idx)].multiply(factor)
-
-    return rhs
