+++ src/pybella/tests/test_unstable_lamb.py
@@ -4,7 +4,8 @@ Unstable Lamb Wave integral test involving vertical Coriolis and Rayleigh BC.
 
 import numpy as np
 from ..flow_solver.physics import hydrostatics
-from ..flow_solver.utils import boundary as bdry
+from ..flow_solver.utils.boundary import cell_boundary as bdry_c
+from ..flow_solver.utils.boundary import rayleigh_boundary as bdry_r
 from ..utils import options as opts
 
 from ..utils.data_structures import DiagnosticState
@@ -269,10 +270,10 @@ def sol_init(Sol, npf, elem, node, th, ud, seeds=None):
             ud.bdry_type[1] = opts.BdryType.RAYLEIGH
 
     if ud.bdry_type[1] == opts.BdryType.RAYLEIGH:
-        ud.tcy, ud.tny = bdry.get_tau_y(ud, elem, node, 0.5)
+        ud.tcy, ud.tny = bdry_r.get_tau_y(ud, elem, node, 0.5)
 
     if ud.rayleigh_forcing:
-        ud.forcing_tcy, ud.forcing_tny = bdry.get_bottom_tau_y(
+        ud.forcing_tcy, ud.forcing_tny = bdry_r.get_bottom_tau_y(
             ud, elem, node, 0.2, cutoff=0.3
         )
 
@@ -354,7 +355,7 @@ def sol_init(Sol, npf, elem, node, th, ud, seeds=None):
 
     npf.p2_nodes[...] = pi_n
 
-    bdry.set_explicit_boundary_data(Sol, elem, ud, th, npf)
+    bdry_c.set_explicit_boundary_data(Sol, elem, ud, th, npf)
 
     if hasattr(ud, "mixed_run"):
         if ud.mixed_run:
